Update install-miniconda.ps1
Change confusing variable name--I expected `python_version` to be something like "2.7" or "3.4", whereas this is the version of the miniconda distributable.  I'm tempted also to change `python_home` to `miniconda_home`, but that is filled below from an environment variable called just "PYTHON" which I'm tempted to also rename to "CONDA" or something.